added requirements for minimal and fetch API, deleted JS crash course link
I think the requirements need to be explicit, since older projects in the repo often lack the minimal API and the linked video does not use it either. Also, the JS crash course was linked on an earlier project and was making the list of resources too long in my opinion.