Enforce .NET SDK version
This is needed to have a reproducible environment for development, but
also in continuous integration.

With `rollForward` set to `latestFeature`, the major and minor versions
must match what's defined in global.json while also being equal to or
greater than the pinned feature band and patch version. So right now,
8.0.300 or any later 8.0.xxx is going to be accepted when using the .NET
SDK.